Update groups-dynamic-membership.md
While working on one of the cases, I have noticed that:
When an iPhone is enrolled in Intune, the OS type displayed in Microsoft Entra changes from "iOS" to "iPhone". This is because of the additional data that is sent to Azure during the enrolment process. When a device is registered in Entra, it shows the OS as "iOS". However, once the device is enrolled in Intune, it sends more detailed information, including the specific device type, which is then reflected as "iPhone" in Entra.

This behaviour has been observed and confirmed through internal testing. The change in OS type helps provide more granular information about the devices being managed, which can be useful for IT administrators when managing and troubleshooting devices.”

Hence Dynamic deivce query for iOS based in device.deviceOSType needed modification as the article currently does not have an information listed regarding the above behavior.
The above behavior is confirmed by the Intune PG. Hence, proposing the update.